+++ src/main/java/miniProject/board/service/ArticleServiceImpl.java
@@ -1,5 +1,6 @@
 package miniProject.board.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import miniProject.board.dto.ArticleDto;
@@ -100,52 +101,44 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public Article update(ArticleDto.Info articleDto, Long articleId, Long memberId) {
+    @Transactional
+    public Article update(ArticleDto.Create articleEditDto, Long articleId, Long memberId) {
         /*
         * 수정 요청을 받은 Article의 id를 받아온다. 존재하지 않는 id이면 수정 불가
         * 수정을 요청한 Member의 세션으로부터 MemberId를 받아온다. 로그아웃 상태인 유저가 게시글을 수정하는 것을 방지
         * 게시글의 작성자 id와 수정을 요청한 Member의 id를 비교. 같은 경우에만 게시글 수정 허용
         */
-        Article article = articleRepository.findById(articleId).orElse(null);
-        if(article == null) {
-            return null;
-        }
+        Article article = articleRepository.findById(articleId)
+                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
 
         Member member = memberService.findMemberDao(memberId);
-        if(member == null) {
-            return null;
+        if(member == null || !member.getId().equals(article.getMember().getId())) {
+            throw new IllegalArgumentException("잘못된 접근입니다.");
         }
 
-        if(Objects.equals(member.getId(), article.getMember().getId())) {
-            article.update(articleDto.getTitle());
-            return articleRepository.save(article);
-        }
+        fileStorageService.updateFile(article.getFilePath(), articleEditDto.getContent());
 
-        return null;
+        article.update(articleEditDto.getTitle());
+        return articleRepository.save(article);
     }
 
     @Override
-    public boolean delete(Long articleId, Long memberId) {
+    public void delete(Long articleId, Long memberId) {
         /*
          *삭제 요청을 받은 Article의 id를 받아온다. 존재하지 않는 id이면 삭제 불가
          *삭제를 요청한 Member의 세션으로부터 MemberId를 받아온다. 로그아웃 상태인 유저가 삭제 요청하는 것을 방지
          *게시글의 작성자 id와 삭제를 요청한 Member의 id를 비교. 같은 경우에만 게시글 삭제 허용
          */
-        Article article = articleRepository.findById(articleId).orElse(null);
-        if(article == null) {
-            return false;
-        }
+        Article article = articleRepository.findById(articleId)
+                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
 
         Member member = memberService.findMemberDao(memberId);
-        if(member == null) {
-            return false;
+        if(member == null || !member.getId().equals(article.getMember().getId())) {
+            throw new IllegalArgumentException("잘못된 접근입니다.");
         }
 
-        if(Objects.equals(member.getId(), article.getMember().getId())) {
-            articleRepository.delete(article);
-            return true;
-        }
+        fileStorageService.deleteFile(article.getFilePath());
 
-        return false;
+        articleRepository.delete(article);
     }
 }
